+++ bin/faucet/src/faucet/mod.rs
@@ -8,7 +8,7 @@ use miden_client::{
     builder::ClientBuilder,
     crypto::RpoRandomCoin,
     keystore::FilesystemKeyStore,
-    note::{NoteError, NoteId, create_p2id_note},
+    note::{NoteError, create_p2id_note},
     rpc::Endpoint,
     transaction::{
         LocalTransactionProver, OutputNote, TransactionId, TransactionProver,
@@ -19,14 +19,12 @@ use rand::{Rng, rng, rngs::StdRng};
 use serde::Serialize;
 use tokio::sync::mpsc::Receiver;
 use tracing::{info, instrument, warn};
-use updates::{ClientUpdater, MintUpdate};
 use url::Url;
 
-use crate::types::{AssetAmount, NoteType};
-
-mod updates;
-
-pub use updates::MintResponseSender;
+use crate::{
+    network::ExplorerUrl,
+    server::{MintRequest, MintResponse, MintResponseSender},
+};
 
 // FAUCET CLIENT
 // ================================================================================================
@@ -56,21 +54,11 @@ impl Serialize for FaucetId {
     }
 }
 
-/// A request for minting to the [`Faucet`].
-pub struct MintRequest {
-    /// Destination account.
-    pub account_id: AccountId,
-    /// Whether to generate a public or private note to hold the minted asset.
-    pub note_type: NoteType,
-    /// The amount to mint.
-    pub asset_amount: AssetAmount,
-}
-
 /// Stores the current faucet state and handles minting requests.
 pub struct Faucet {
     id: FaucetId,
     decimals: u8,
-    client: Client,
+    client: Client<FilesystemKeyStore<StdRng>>,
     tx_prover: Arc<dyn TransactionProver>,
 }
 
@@ -156,37 +144,39 @@ impl Faucet {
 
     /// Runs the faucet minting process until the request source is closed, or it encounters a fatal
     /// error.
+    ///
+    /// It receives new minting requests and handles them in batches. For each request, it builds a
+    /// minting note that is included in a new transaction. For each request, sends the
+    /// resulting `MintResponse` through the response sender.
     pub async fn run(
         mut self,
         mut requests: Receiver<(MintRequest, MintResponseSender)>,
     ) -> anyhow::Result<()> {
         let mut buffer = Vec::new();
         let limit = 256; // also limited by the queue size `REQUESTS_QUEUE_SIZE`
+        let explorer_url = ExplorerUrl::from_network_id(self.id.network_id);
 
         while requests.recv_many(&mut buffer, limit).await > 0 {
-            // Skip requests where the user no longer cares about the result.
-            let (requests, response_senders): (Vec<MintRequest>, Vec<MintResponseSender>) = buffer
-                .drain(..)
-                .filter(|(request, response_sender)| {
-                    if response_sender.is_closed() {
-                        tracing::info!(request.account_id=%request.account_id, "request cancelled");
-                        false
-                    } else {
-                        true
-                    }
-                })
-                .unzip();
+            let (requests, response_senders): (Vec<MintRequest>, Vec<MintResponseSender>) =
+                buffer.drain(..).unzip();
 
-            let updater = ClientUpdater::new(response_senders);
+            let mut rng = {
+                let auth_seed: [u64; 4] = rng().random();
+                let rng_seed = Word::from(auth_seed.map(Felt::new));
+                RpoRandomCoin::new(rng_seed)
+            };
+            let notes = build_p2id_notes(self.id, self.decimals, &requests, &mut rng)?;
+            let note_ids = notes.iter().map(OutputNote::id).collect::<Vec<_>>();
+            let tx_id = self.create_transaction(notes).await?;
 
-            let (tx_id, note_ids) = self.handle_request_batch(&requests, &updater).await?;
-
-            for note_id in note_ids {
-                updater
-                    .send_updates(MintUpdate::Minted(note_id, tx_id, self.id.network_id))
-                    .await;
+            for (sender, note_id) in response_senders.into_iter().zip(note_ids) {
+                // Ignore errors if the request was dropped.
+                let _ = sender.send(MintResponse {
+                    tx_id,
+                    note_id,
+                    explorer_url: explorer_url.clone(),
+                });
             }
-
             self.client.sync_state().await?;
         }
 
@@ -195,32 +185,19 @@ impl Faucet {
         Ok(())
     }
 
-    /// Fully handles a batch of requests to create and submit a transaction.
-    ///
-    /// For each mint request, a mint note is built. Then, with these notes, a transaction is
-    /// created, executed, and submitted using the local miden-client. This results in submitting
-    /// the transaction to the node and updating the local db to track the created notes.
-    async fn handle_request_batch(
+    /// Creates a transaction with the given notes, executes it, proves it, and submits using the
+    /// local miden-client. This results in submitting the transaction to the node and updating the
+    /// local db to track the created notes.
+    async fn create_transaction(
         &mut self,
-        requests: &[MintRequest],
-        updater: &ClientUpdater,
-    ) -> Result<(TransactionId, Vec<NoteId>), ClientError> {
-        let mut rng = {
-            let auth_seed: [u64; 4] = rng().random();
-            let rng_seed = Word::from(auth_seed.map(Felt::new));
-            RpoRandomCoin::new(rng_seed)
-        };
-
-        // Build the notes
-        let notes = build_p2id_notes(self.id, self.decimals, requests, &mut rng)?;
-        let note_ids = notes.iter().map(OutputNote::id).collect();
+        notes: Vec<OutputNote>,
+    ) -> Result<TransactionId, ClientError> {
+        // Build the transaction
         let tx = TransactionRequestBuilder::new().own_output_notes(notes).build()?;
-        updater.send_updates(MintUpdate::Built).await;
 
         // Execute the transaction
         let tx_result = self.client.new_transaction(self.id.account_id, tx).await?;
         let tx_id = tx_result.executed_transaction().id();
-        updater.send_updates(MintUpdate::Executed).await;
 
         // Prove and submit the transaction
         let prover_failed = self
@@ -232,9 +209,8 @@ impl Faucet {
             warn!("Failed to prove transaction with remote prover, falling back to local prover");
             self.client.submit_transaction(tx_result).await?;
         }
-        updater.send_updates(MintUpdate::Submitted).await;
 
-        Ok((tx_id, note_ids))
+        Ok(tx_id)
     }
 
     /// Returns the id of the faucet account.
@@ -275,113 +251,4 @@ fn build_p2id_notes(
         notes.push(OutputNote::Full(note));
     }
     Ok(notes)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{str::FromStr, time::Duration};
-
-    use miden_client::{
-        account::{AccountBuilder, AccountStorageMode, AccountType, component::RpoFalcon512},
-        asset::TokenSymbol,
-        auth::AuthSecretKey,
-        crypto::SecretKey,
-        note::BlockNumber,
-        rpc::{NodeRpcClient, TonicRpcClient},
-        store::TransactionFilter,
-    };
-    use rand::Rng;
-    use tokio::time::{Instant, sleep};
-    use url::Url;
-
-    use super::*;
-    use crate::{testing::stub_rpc_api::serve_stub, types::AssetOptions};
-
-    // This test ensures that the faucet can create a transaction that outputs a batch of notes.
-    #[allow(clippy::cast_sign_loss)]
-    #[tokio::test]
-    async fn faucet_batches_requests() {
-        let stub_node_url = Url::from_str("http://localhost:50052").unwrap();
-
-        // Start the stub node
-        tokio::spawn({
-            let stub_node_url = stub_node_url.clone();
-            async move { serve_stub(&stub_node_url).await.unwrap() }
-        });
-        // Wait for the stub node to serve requests
-        let rpc_client =
-            TonicRpcClient::new(&Endpoint::try_from(stub_node_url.as_str()).unwrap(), 1000);
-        let start = Instant::now();
-        while rpc_client
-            .get_block_header_by_number(Some(BlockNumber::GENESIS), false)
-            .await
-            .is_err()
-        {
-            sleep(Duration::from_millis(100)).await;
-            assert!(start.elapsed() < Duration::from_secs(5), "stub node took too long to start");
-        }
-
-        // Create the faucet
-        let mut faucet = {
-            let secret = SecretKey::new();
-            let symbol = TokenSymbol::try_from("MIDEN").unwrap();
-            let decimals = 2;
-            let max_supply = Felt::try_from(1_000_000_000_000u64).unwrap();
-            let (account, account_seed) = AccountBuilder::new(rng().random())
-                .account_type(AccountType::FungibleFaucet)
-                .storage_mode(AccountStorageMode::Public)
-                .with_component(BasicFungibleFaucet::new(symbol, decimals, max_supply).unwrap())
-                .with_auth_component(RpoFalcon512::new(secret.public_key()))
-                .build()
-                .unwrap();
-            let account_file = AccountFile::new(
-                account,
-                Some(account_seed),
-                vec![AuthSecretKey::RpoFalcon512(secret)],
-            );
-
-            Faucet::load(
-                PathBuf::from("faucet_client_store.sqlite3"),
-                NetworkId::Testnet,
-                account_file,
-                &stub_node_url,
-                Duration::from_secs(10),
-                None,
-            )
-            .await
-            .unwrap()
-        };
-
-        // Create a set of mint requests
-        let num_requests = 5;
-        let requests = (0..num_requests)
-            .map(|i| {
-                let account_id = (i as u128).try_into().unwrap();
-                MintRequest {
-                    account_id,
-                    asset_amount: AssetOptions::new(vec![100]).unwrap().validate(100).unwrap(),
-                    note_type: NoteType::Public,
-                }
-            })
-            .collect::<Vec<_>>();
-
-        let (tx_id, note_ids) = faucet
-            .handle_request_batch(&requests, &ClientUpdater::new(vec![]))
-            .await
-            .unwrap();
-
-        let tx = faucet
-            .client
-            .get_transactions(TransactionFilter::Ids(vec![tx_id]))
-            .await
-            .unwrap()
-            .first()
-            .unwrap()
-            .clone();
-        assert_eq!(tx.details.output_notes.num_notes(), num_requests as usize);
-        assert_eq!(
-            tx.details.output_notes.iter().map(OutputNote::id).collect::<Vec<_>>(),
-            note_ids
-        );
-    }
 }