+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Changes
 
-- Integrated miden-client (#11).
-- Added `/get_note` endpoint (#19).
-- Redesigned the home frontend (#20).
+- Integrated miden-client ([#11](https://github.com/0xMiden/miden-faucet/pull/11)).
+- Added `/get_note` endpoint ([#19](https://github.com/0xMiden/miden-faucet/pull/19)).
+- Redesigned the home frontend ([#20](https://github.com/0xMiden/miden-faucet/pull/20)).
+- Redesigned the tokens request flows ([#25](https://github.com/0xMiden/miden-faucet/pull/25)).
 
 ## 0.10.0 (2025-07-10)